+++ src/pages/Items.jsx
@@ -1,53 +1,13 @@
-import { useState, useEffect } from "react";
-import { getProductsList } from "../services/panda-market-api";
-import ProductsList from "../layout/ProductsList";
+import AllProducts from "../components/Product/AllProducts";
+import BestProducts from "../components/Product/BestProducts";
 import "../styles/page-items/items.css";
 
-const Items = (props) => {
-  const [AllItems, setAllItems] = useState([]);
-  const [RecItems, setRecItems] = useState([]);
-  const [search, setSearch] = useState("");
-  const [order, setOrder] = useState("recent");
-
-  const handleLoad = async (options) => {
-    const { list } = await getProductsList(options);
-    if (options.pageSize === 4) {
-      setRecItems(list);
-    } else {
-      setAllItems(list);
-    }
-  };
-
-  const handleSearchSubmit = (e) => {
-    e.preventDefault();
-    setSearch(e.target["search"].value);
-  };
-
-  const handleBestClick = () => setOrder("favorite");
-  const handleNewestClick = () => setOrder("recent");
-
-  useEffect(() => {
-    handleLoad({ pageSize: 4, orderBy: "favorite" });
-    handleLoad({ pageSize: 10, keyword: search, orderBy: order });
-  }, [search, order]);
-
+const Items = () => {
   return (
     <main className='page-items'>
       <div className='container'>
-        <ProductsList list={RecItems} imageSize='large' countSize='small' productManagement={false}>
-          베스트 상품
-        </ProductsList>
-        <ProductsList
-          list={AllItems}
-          imageSize='middle'
-          countSize='small'
-          productManagement={true}
-          onSubmit={handleSearchSubmit}
-          onBest={handleBestClick}
-          onNewest={handleNewestClick}
-        >
-          전체 상품
-        </ProductsList>
+        <BestProducts />
+        <AllProducts />
       </div>
     </main>
   );
